archsim: Make some fixes to GenC API functions

+++ archsim/inc/gensim/gensim_processor_api.h
@@ -95,7 +95,7 @@ static inline uint64_t bitcast_double_u64(double x)
 extern "C" {
 
 	uint32_t genc_bswap32(uint32_t);
-	uint32_t genc_bswap64(uint64_t);
+	uint64_t genc_bswap64(uint64_t);
 
 	uint32_t genc_adc_flags(uint32_t lhs, uint32_t rhs, uint32_t carry_in);
 	uint32_t genc_adc64_flags(uint64_t lhs, uint64_t rhs, uint32_t carry_in);
